move `from_random_normal` to external function
